+++ encode/src/main/scala/org/broadinstitute/monster/etl/encode/transforms/package.scala
@@ -1,0 +1,10 @@
+package org.broadinstitute.monster.etl.encode
+
+import com.spotify.scio.values.SCollection
+import io.circe.JsonObject
+
+package object transforms {
+
+  /** Convenience alias for transformations that operate on JSON. */
+  type JsonPipe = SCollection[JsonObject] => SCollection[JsonObject]
+}
